Move RedisEC to separate file
